python scripts -- correction on the re-sampling of cross-section through meshes (ray trace) to allow for what[extract] to be close to triangle edges.
git-svn-id: http://svn.opentelemac.org/svn/opentelemac/trunk@2662 21fd64ea-cbdc-4a4a-8fb2-7ac88764eefb

Former-commit-id: 8f6394cbcb7bab26abe55e228b8654eb67ae627a

+++ pytel/parsers/parserStrings.py
@@ -50,7 +50,7 @@ def parseArrayPaires(s):
          if procd:
             p.append(float(procd.group('number')))
          elif proci:
-            p.append(float(proci.group('number')))
+            p.append(int(proci.group('number')))
          else:
             print '... could not parse the array: ' + s
             sys.exit()
